cast changeset errors even if they have other metadata attached
I've noticed this for a while but it's not a big problem... but I think the reason why a lot of our changeset error codes get translated to `"unknown_error"` is because the keyword list in the changeset can include other properties. It would be nice to probably also send those to the frontend really -- for example, if the error is "Name is limited to 100 characters", returning the max characters in the payload is helpful for localization e.g. `"code": "max_length", "message": "Name is limited to 100 characters", "validation": { "max_length": 100 }`

note: untested, just wrote in GitHub as a proposal

Signed-off-by: Matt Miller <matt.miller@smartrent.com>